{{cookiecutter.project_slug}}/requirements.txt
-e .
PasteDeploy>=1.5,<1.6
gunicorn
